Fix Doppler effect in background
Decided that maybe its better to rewrite the Doppler effect in terms of relative motion rather than just for a stationary observer? Not sure. But, the limit for v<<c requires relative motion at least in the relativistic case. And that would make it easier to compare them but makes the doppler effect paragraph kinda draggish.